refactor templater into individual package
